pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
pydantic-settings = "^2.6.1"
gunicorn = "^22.0.0"
greenlet = "^3.0.3"
loguru = "^0.7.2"
motor = "^3.6.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
isort = "^5.13.2"
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"
black = "^24.8.0"
mypy = "^1.11.2"
pytest-mock = "^3.14.0"
mongomock = "^4.3.0"
mongomock-motor = "^0.0.34"
python-multipart = "^0.0.19"

[tool.isort]
profile = "black"
line_length = 90
multi_line_output = 3
skip_gitignore = true
skip_glob = ["**/alembic/*", "**/config/*"]
src_paths = ["./src"]


[tool.mypy]
files = "src"
exclude = "(alembic|venv|config.py|tests)"
ignore_missing_imports = true
disallow_untyped_calls = false
disallow_untyped_defs = true
pretty = true
strict_optional = true
follow_imports = "silent"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
